# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Provision Civils - Construction Job Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      responses:
        "200":
          description: Employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"

    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found

    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /employees/{id}/banking:
    get:
      operationId: getEmployeeBanking
      tags: [employees]
      summary: Get employee banking details (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Banking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeBanking"
        "404":
          description: Not found

    put:
      operationId: upsertEmployeeBanking
      tags: [employees]
      summary: Create or update employee banking details (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeBankingInput"
      responses:
        "200":
          description: Banking details saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeBanking"
...
components:
  schemas:
...
    Employee:
      type: object
      required: [id, name, email, role, employmentStatus, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, project_manager, supervisor, worker]
        phone:
          type: ["string", "null"]
        employeeNumber:
          type: ["string", "null"]
        clockNumber:
          type: ["string", "null"]
        idNumber:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        homeAddress:
          type: ["string", "null"]
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactNumber:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        supervisorId:
          type: ["integer", "null"]
        employmentStartDate:
          type: ["string", "null"]
        employmentStatus:
          type: string
          enum: [active, suspended, resigned, dismissed]
        payrollType:
          type: ["string", "null"]
          enum: [hourly, piece_work, "null"]
        hourlyRate:
          type: ["string", "null"]
        meterRate:
          type: ["string", "null"]
        createdAt:
          type: string
    EmployeeInput:
      type: object
      required: [name, email, role, password]
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, project_manager, supervisor, worker]
        phone:
          type: string
        password:
          type: string
        employeeNumber:
          type: string
        clockNumber:
          type: string
        idNumber:
          type: string
        dateOfBirth:
          type: string
        homeAddress:
          type: string
        emergencyContactName:
          type: string
        emergencyContactNumber:
          type: string
        jobTitle:
          type: string
        department:
          type: string
        supervisorId:
          type: integer
        employmentStartDate:
          type: string
        employmentStatus:
          type: string
          enum: [active, suspended, resigned, dismissed]
        payrollType:
          type: string
          enum: [hourly, piece_work]
        hourlyRate:
          type: number
        meterRate:
          type: number
    EmployeeUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, project_manager, supervisor, worker]
        phone:
          type: string
        employeeNumber:
          type: string
        clockNumber:
          type: string
        idNumber:
          type: string
        dateOfBirth:
          type: string
        homeAddress:
          type: string
        emergencyContactName:
          type: string
        emergencyContactNumber:
          type: string
        jobTitle:
          type: string
        department:
          type: string
        supervisorId:
          type: integer
        employmentStartDate:
          type: string
        employmentStatus:
          type: string
          enum: [active, suspended, resigned, dismissed]
        payrollType:
          type: string
          enum: [hourly, piece_work]
        hourlyRate:
          type: number
        meterRate:
          type: number
    EmployeeBanking:
      type: object
      required: [id, userId, bankName, accountHolder, accountNumber, branchCode, accountType]
      properties:
        id:
          type: integer
        userId:
          type: integer
        bankName:
          type: string
        accountHolder:
          type: string
        accountNumber:
          type: string
        branchCode:
          type: string
        accountType:
          type: string
          enum: [cheque, savings, transmission]
    EmployeeBankingInput:
      type: object
      required: [bankName, accountHolder, accountNumber, branchCode, accountType]
      properties:
        bankName:
          type: string
        accountHolder:
          type: string
        accountNumber:
          type: string
        branchCode:
          type: string
        accountType:
          type: string
          enum: [cheque, savings, transmission]
...